Refactor parcel id code checks in to a function
This change moves the is_null and preg_match calls, which check the
validity of the parcel id in to a function. This allows for more
sophisticated checking later, should it be required, or simplified
checking, which we'll see in the next commit. Would be handy to know if
PhpStorm can move code to a function too.